+++ src/BlinkList.h
@@ -10,9 +10,9 @@ private:
         uint8_t gpio;
 bool enabled;
 bool state;
-unsigned long on_tact;
-unsigned long off_tact;
-unsigned long tact_counter:
+unsigned int on_tact;
+unsigned int off_tact;
+unsigned int tact_counter:
         Node* next;
     };
 
@@ -23,7 +23,7 @@ public:
     BlinkList();                 // Constructor
     ~BlinkList();                // Destructor
 
-      Node* configure(uint8_t gpio, unsigned long t_on_ms, unsigned long t_off_ms     // Add a new node to the end and configure  it
+      Node* configure(uint8_t gpio, unsigned int t_on_ms, unsigned int t_off_ms     // Add a new node to the end and configure  it
     Node* search(uint8_t gpio);      // Search for a value
     bool remove(uint8_t gpio);      // Remove a node by value
     void print();                // Print all elements (for debug)
