+++ src/bnode.h
@@ -8,10 +8,10 @@ public:
     uint8_t gpio;
     bool enabled;
     bool state;
-    unsigned long tact_on;
-    unsigned long tact_off;
-    unsigned long tact_counter;
-    BNode* next;
+    unsigned int tact_on;
+    unsigned int tact_off;
+    unsigned int tact_counter;
+    BNode *next;
 
     BNode(uint8_t gpio);
 };
